Remove oneline argument from swiftString
This should fix the issue where, if you start a string which a single
non-matched quote, the string and interpolated string isn't highlighted
until the quote was closed. This seems like a bad idea since strings
only function when they are on a single line and this example:

"foo
" bar

highlights like they can spread over multiple lines. Where `bar` is
highlighted like a keyword and not part of an unclosed string. This is
also how ruby.vim does this so I assume this is the way to do it.

This fixes #28