Move comment pattern, update division operator
This fixes an issue where the beginning /* in a C-style block comment
was interpreted as an operator. I read somewhere in syntax.txt (I think)
that the order rules appear in the syntax definition file affects the
interpretation. This is one of those times. The comment block needs to
be after the operator definitions.