Remove swiftLiteralWrapper for now.
I could get this to work. The failed use case is this:

baseURL = "http://www.oaklandpostonline.com/search/?q&tarticle&c[]blogs*"

The start pattern matches the `=` right after baseURL and ends on `[]`
in the URL. I couldn't get this to ignore the swiftString type or ignore
if there was a quote right after the URL. Mainly everything is horrible.

Closes #37