Improve highlighting of `#available` condition
Conforming to Xcode, the syntax file now highlights platforms contained
within the `#available` condition.

It also now removes highlighting numbers and the "*" (i.e, symbol for
future platforms) that are contained within the `#available` condition
since they are not literal numbers or operator (thanks to @acclissold
for the idea)

Additionally, add tests to example.swift to demonstrate highlighting of
them.

References #56