Fix comment and string indentation issue
This updates the previously dumb NumberOfMatches function to take the
index of the line, and make sure the match at that line shouldn't be
ignored. It should be ignored if the character is part of a string or a
comment. Without this change those characters could be taken into
account when indenting subsequent lines.

Closes #65