Force vim to sync 100 lines
This fixes a bug where if you had a very long documentation comment, and
you launched vim and it resumed with only being able to see a portion of
that comment, the comment would not be highlighted correctly. By default
ruby syncs 500 lines so we could probably up this number if needed, but
this should be fine for now.

Closes #49