Fix ] indent issue
Previously when you typed, or tried to reindent a ] character, it didn't
work. This was because the brackets weren't escaped in search pair. This
also fixes auto indenting when typing the close square bracket. This is
now special cased because search pair returns no matches right after
typing the bracket.