Fix multiline paren indenting
This fixes the python-style indenting that was previously done. If you
had a codeblock like this:

```swift
let foo = Thing(arg: arg,
                arg2: arg2)
```

The indent of the second line would be the opening paren +1. This has
changed to:

```swift
let foo = Thing(arg: arg,
    arg2: arg2)
```

Which is the indent of the opening paren line +1 'shiftwidth'